+++ tsconfig.json
@@ -1,11 +1,11 @@
 {
     "compilerOptions": {
-        "outDir": "./admin",
+        "rootDir": "./js",
+        "outDir": "./admin/",
         "allowJs": false,
         "target": "ES2017",
+        "module": "commonjs",
+        "sourceMap": true,
         "allowSyntheticDefaultImports": true
-    },
-    "include": [
-        "./js/wp2static-admin.ts"
-    ]
+    }
 }
