+++ webpack.config.js
@@ -1,0 +1,24 @@
+var path = require('path');
+var webpack = require('webpack');
+
+module.exports = {
+    entry: {
+        main: './js/wp2static-admin.ts',
+    },
+    resolve: {
+        extensions: [".webpack.js", ".web.js", ".js", ".ts"]
+    },
+    output: {
+        //publicPath: "/admin/",
+        path: path.resolve(__dirname, 'admin'),
+        filename: 'wp2static-admin.js'
+    },
+    module: {
+        rules: [
+            {
+                test: /\.ts$/,
+                loader: 'ts-loader'
+            }
+        ]
+    }
+};
